+++ lirical-cli/src/main/java/org/monarchinitiative/lirical/cli/cmd/AbstractPrioritizeCommand.java
@@ -11,12 +11,10 @@ import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
 
 import java.nio.file.Path;
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
+import java.time.LocalDateTime;
 import java.time.Period;
 import java.time.format.DateTimeParseException;
 import java.util.*;
-import java.util.function.Function;
 
 /**
  * This is a common superclass for {@link YamlCommand}, {@link PhenopacketCommand}, and {@link PrioritizeCommand}.
@@ -39,9 +37,13 @@ abstract class AbstractPrioritizeCommand extends BaseLiricalCommand {
         public Path outdir = Path.of("");
 
         @CommandLine.Option(names = {"-f", "--output-format"},
-                paramLabel = "{html,tsv}",
-                description = "Comma separated list of output formats to use for writing the results (default: ${DEFAULT-VALUE}).")
-        public String outputFormats = "html";
+                arity = "0..*",
+                description = {
+                "An output format to use for writing the results, can be provided multiple times.",
+                        "Choose from {${COMPLETION-CANDIDATES}}",
+                        "Default: ${DEFAULT-VALUE}"
+                })
+        public Set<OutputFormat> outputFormats = Set.of(OutputFormat.HTML);
         /**
          * Prefix of the output file. For instance, if the user enters {@code -x sample1} and an HTML file is output,
          * the name of the HTML file will be {@code sample1.html}. If a TSV file is output, the name of the file will
@@ -99,7 +101,7 @@ abstract class AbstractPrioritizeCommand extends BaseLiricalCommand {
                 .setTranscriptDatabase(runConfiguration.transcriptDb.toString())
                 .setLiricalPath(dataSection.liricalDataDirectory.toAbsolutePath().toString())
                 .setExomiserPath(dataSection.exomiserDatabase == null ? "" : dataSection.exomiserDatabase.toAbsolutePath().toString())
-                .setAnalysisDate(getTodaysDate())
+                .setAnalysisDate(LocalDateTime.now().toString())
                 .setSampleName(analysisData.sampleId())
                 .setnGoodQualityVariants(filteringStats.nGoodQualityVariants())
                 .setnFilteredVariants(filteringStats.nFilteredVariants())
@@ -108,9 +110,13 @@ abstract class AbstractPrioritizeCommand extends BaseLiricalCommand {
                 .build();
 
         OutputOptions outputOptions = createOutputOptions();
-        lirical.analysisResultsWriterFactory()
-                .getWriter(analysisData, results, metadata)
-                .process(outputOptions);
+        AnalysisResultWriterFactory factory = lirical.analysisResultsWriterFactory();
+
+        for (OutputFormat fmt : output.outputFormats) {
+            Optional<AnalysisResultsWriter> writer = factory.getWriter(fmt);
+            if (writer.isPresent())
+                writer.get().process(analysisData, results, metadata, outputOptions);
+        }
 
         reportElapsedTime(start, System.currentTimeMillis());
         return 0;
@@ -140,9 +146,8 @@ abstract class AbstractPrioritizeCommand extends BaseLiricalCommand {
     protected OutputOptions createOutputOptions() {
         LrThreshold lrThreshold = output.lrThreshold == null ? LrThreshold.notInitialized() : LrThreshold.setToUserDefinedThreshold(output.lrThreshold);
         MinDiagnosisCount minDiagnosisCount = output.minDifferentialsToShow == null ? MinDiagnosisCount.notInitialized() : MinDiagnosisCount.setToUserDefinedMinCount(output.minDifferentialsToShow);
-        List<OutputFormat> outputFormats = parseOutputFormats(output.outputFormats);
         return new OutputOptions(lrThreshold, minDiagnosisCount, runConfiguration.pathogenicityThreshold,
-                output.displayAllVariants, output.outdir, output.outfilePrefix, outputFormats);
+                output.displayAllVariants, output.outdir, output.outfilePrefix);
     }
 
     protected static Age parseAge(String age) {
@@ -157,34 +162,6 @@ abstract class AbstractPrioritizeCommand extends BaseLiricalCommand {
         } catch (DateTimeParseException e) {
             throw new LiricalRuntimeException("Unable to parse age '" + age + "': " + e.getMessage(), e);
         }
-    }
-
-    /**
-     * @return a string with today's date in the format yyyy/MM/dd.
-     */
-    private static String getTodaysDate() {
-        DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd");
-        Date date = new Date();
-        return dateFormat.format(date);
-    }
-
-    private List<OutputFormat> parseOutputFormats(String outputFormats) {
-        return Arrays.stream(outputFormats.split(","))
-                .map(String::trim)
-                .map(toOutputFormat())
-                .flatMap(Optional::stream)
-                .toList();
-    }
-
-    private static Function<String, Optional<OutputFormat>> toOutputFormat() {
-        return payload -> switch (payload.toUpperCase()) {
-            case "HTML" -> Optional.of(OutputFormat.HTML);
-            case "TSV" -> Optional.of(OutputFormat.TSV);
-            default -> {
-                LOGGER.warn("Unknown output format {}", payload);
-                yield Optional.empty();
-            }
-        };
     }
 
 }