+++ lirical-core/src/main/java/org/monarchinitiative/lirical/core/output/AnalysisResultsWriter.java
@@ -2,51 +2,23 @@ package org.monarchinitiative.lirical.core.output;
 
 import org.monarchinitiative.lirical.core.analysis.AnalysisData;
 import org.monarchinitiative.lirical.core.analysis.AnalysisResults;
-import org.monarchinitiative.phenol.annotations.formats.hpo.HpoDiseases;
-import org.monarchinitiative.phenol.ontology.data.Ontology;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import java.io.IOException;
 
-public class AnalysisResultsWriter {
+/**
+ * The implementors can store results of LIRICAL analysis in some {@link OutputFormat}.
+ */
+public interface AnalysisResultsWriter {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(AnalysisResultsWriter.class);
-
-    private final Ontology hpo;
-    private final HpoDiseases diseases;
-    private final AnalysisData analysisData;
-    private final AnalysisResults analysisResults;
-    private final AnalysisResultsMetadata metadata;
-
-
-    AnalysisResultsWriter(Ontology hpo,
-                          HpoDiseases diseases,
-                          AnalysisData analysisData,
-                          AnalysisResults analysisResults,
-                          AnalysisResultsMetadata metadata) {
-        this.hpo = hpo;
-        this.diseases = diseases;
-        this.analysisData = analysisData;
-        this.analysisResults = analysisResults;
-        this.metadata = metadata;
-    }
-
-    public void process(OutputOptions options) {
-        for (OutputFormat format : options.outputFormats()) {
-            templateForFormat(format, options)
-                    .ifPresent(LiricalTemplate::outputFile);
-        }
-    }
-
-    private Optional<LiricalTemplate> templateForFormat(OutputFormat format, OutputOptions options) {
-        // TODO - finalize template creation
-        return switch (format) {
-            case TSV -> Optional.of(new TsvTemplate(hpo, diseases, analysisData, analysisResults, metadata, options));
-            case HTML -> Optional.of(new HtmlTemplate(hpo, diseases, analysisData, analysisResults, metadata, options, List.of(), Set.of()));
-        };
-    }
+    /**
+     * Write the provided {@link AnalysisData}, {@link AnalysisResults} and {@link AnalysisResultsMetadata} into
+     * some {@link OutputFormat} as guided by {@link OutputOptions}.
+     *
+     * @throws IOException in case of I/O errors
+     */
+    void process(AnalysisData analysisData,
+                 AnalysisResults analysisResults,
+                 AnalysisResultsMetadata metadata,
+                 OutputOptions outputOptions) throws IOException;
 
 }